FileSystemExplorer: Add missing import
When running on the example, qmllint would complain that it doesn't know
about Qt.RightEdge (and BottomEdge). The reason for that is that the
global Qt namespace is only exposed by the global Qt object from
builtins. Controls doesn't actually expose that one, so we need to
import QtQml (QML would work, too, but that's mostly internal).

Pick-to: 6.10
Fixes: QTBUG-138171
Change-Id: I28004e1d36ae6d07a1a01a72ecf716b17ca5b9ec
Reviewed-by: Dennis Oberst <dennis.oberst@qt.io>
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>