qmllint: don't warn about fall-throughs where they don't happen
Iterate through the statements of the case-statements and don't warn
about a fall-through if the code path of the case-statements return or
break or throw before reaching the end of the case-block.

Amends 65804aaa928bbf6467e22da4bd35fcae1373af2d that introduced the
warning and only didn't check recursively for break, return or throw
statements inside of case blocks.

Pick-to: 6.10
Fixes: QTBUG-138532
Change-Id: I545898ebd52a4ec8f0ea58ca4ab5519e7137b6b0
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
Reviewed-by: Lars Schmertmann <SmallLars@t-online.de>