QQmlPropertyCacheCreator: let PropertyCache handle override instead of Resolver
Existing handling of `final` using Resolver seems to be redundant
unless there are some corner cases, not covered by our tests.

For maintainability purposes it's arguably better to keep handling
of override logic in one place and inside PropertyCache,
which is intended to be responsible for it (see handleOverride).
Therefore adjusting PropertyCache::appendProperty to propagate
OverrideResult to the upstream calls and use it in
PropertyCacheCreator<>::tryDeriveCacheFrom to throw an error in
cases of invalid property override.

Change-Id: Iec8da8d5bd2979a4639c27d916b9444b37ad0b7a
Task-number: QTBUG-98320
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>