FluentWinUI3: Improve SearchField indicators
The SearchField indicators draw a background when hovered/pressed which
was not showing on light mode because it was using a white color.
Instead of drawing the background ourselves use image assets from the
Spinbox indicator-s as they share the same style.
In addition, use a palette color for the indicator icons as well as
change the opacity when it is pressed.

Pick-to: 6.10
Change-Id: I4f4413cb408d80d2167ca52b98cefdf4b51f25f2
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>