QQ4A: Add QtSignalQueue class to manage queued signal connections
This class will be used in both QtQuickView and QtQuickViewContent, to
keep track of queued signal connections and eventually process them,
connecting them to the QML view.

This class also introduces an internal class to represent the data
required to make a successful signal connection, to assist with storing
the data.

Task-number: QTBUG-137057
Change-Id: I798dbf0c0b377f2bf07874b17b83d493aa4905d9
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>