Docs: split the SQL section out from the C++ model documentation
Integration of Qt Quick's item views with SQL databases is a significant
enough topic to deserve it's own page, and the overlap with exposing C++
data structures to QML is minimal. So move the section into a separate
page, and adjust the section levels accordingly.

Link to the Model/View Programming documentation for creating an
editable model. It's more complete and authoritative documentation
than the somewhat brief explanation of QAIM::setData in the cppmodels
documentation page.

Pick-to: 6.10
Change-Id: I52b759d2bd28ea6227e69a33653ba608fe530b20
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>