QQmlPropertyCacheCreator: add tryDeriveCacheFrom(objIR, baseTypePropCache)
This commit aims to break monolithic logic of property cache creation
into smaller isolated chunks by extracting and exposing a method
responsible for creation of cache based on the object's IR and
propertyCache of base type.

Currently there is a lot of coupled logic happening behind the
buildMetaObjectsIncrementally, which is extremely hard to understand and
especially hard to test in isolation.
It requires a lot of pre-setup of different components to make things work.

Within the scope of QTBUG-98320, however, all what's needed to be verified
is that given objectIR and property cache of Base Type we create/derive a
correct property cache for this object.
a.k.a. (objectIR, BaseTypePropertyCache) -> objectPropertyCache

The logic in `createMetaObject`, which is a protected method is overly
complicated by itself and depends on the state of the PropertyCacheCreator,
which is changing during buildMetaObjectsIncrementally.
However this "derivation" can be easily extracted and isolated, hence
better tested.

Some tests will follow in the further commits.

Task-number: QTBUG-98320
Change-Id: I332591b0f210efb95d5c096cc75e77d837d9ab94
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>