QQmlPropertyCacheCreator: let PropertyCache handle override instead of Resolver (2/2)
Existing handling of `final` using Resolver seems to be redundant
unless there are some corner cases, not covered by our tests.

For maintainability purposes it's arguably better to keep handling
of override logic in one place and inside PropertyCache,
which is intended to be responsible for it (see handleOverride).
Therefore adjusting PropertyCache::appendAlias to propagate
OverrideResult to the upstream calls and use it in
QQmlPropertyCacheAliasCreator<>::appendAliasToPropertyCache
to throw an error in cases of invalid property override.

Change-Id: I7cabd42197b0829dd7fc2965868c5e5287169a40
Task-number: QTBUG-98320
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>