Backport some changes from QmlTreeModel to QmlTableModel
QmlTreeModel was recently added and in practice it is an extension
to QmlTableModel. Many functions were initially copied from
QmlTableModel and then modified. Some changes are also applicable to the
QmlTableModel, e.g. the usage of auto, Latin1StringView, enums instead
of bools as well as some asserts and white spaces.

Also some names have been made consistent and whitespace was corrected.

This is the 2nd patch in an attempt to make QQmlTreeModel and
QQmlTableModel more similar and eventually merge common parts
in a common abstract class.

Task-number: QTBUG-138703
Pick-to: 6.10
Change-Id: Ic6ce7a5e1c644ab99bb998235fcd8a7087a124e4
Reviewed-by: Mate Barany <mate.barany@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>