Make QmlTreeModel more similar to QmlTableModel
Some things are solved better in QmlTableModel and they are ported to
QmlTreeModel. Some debug output that is present in QmlTableModel was
introduced in QmlTreeModel.

Whitespaces were corrected.

This is the 5th patch in an attempt to make QQmlTreeModel and
QQmlTableModel more similar and eventually merge common parts
in a common abstract class.

Task-number: BUG-138703
Pick-to: 6.10
Change-Id: I62bd1b484951cf619af68e0b3aef8c056a88d407
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>