Quick: Correctly expose more types to QML
In order for QML tooling (and in some cases, the runtime [1]) to work
correctly,  meta-methods need to use the fully qualified type. Moreover,
we need to make sure that types which are referenced are exposet to QML,
even if only as anonymous types.
Switching signals and  slots to fully qualified types risks breaking old
style connects, so this is only applied to private headers.

This reduces the warnings reported by the Quick_qmllint_module target
quite a bit.

[1] The runtime often has the meta-type available, which is
generally enough. Tooling however will not be able to identify the type
and provide e.g. proper auto-completion.

Pick-to: 6.10
Change-Id: I706c25d8fdeb79ca3d2488ce1d2e66f79818cfbb
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>