Add insertRow() to QQmlTreeModel
QQmlTreeModel lacks an API to insert a row into the parent's list
of children at a specified position. Working around this by
using removeRow() and appendRow() is possible but inconvenient.

This patch introduces insertRow() as an additional function.

Similar to appendRow(), it adds a row to a parent but at a
selectable position. An overload to insert at the top
level exists, as with appendRow().

[ChangeLog][Qt labs QML Models] Added a new method to QQmlTreeModel
that inserts a node into the list of parent's children at the given
index.

Fixes: QTBUG-138704
Change-Id: I2de974b208f6aa8b02c84bf7b2bdd7d79ca84f4c
Reviewed-by: Matthias Rauter <matthias.rauter@qt.io>
Reviewed-by: Mate Barany <mate.barany@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>