QQmlPropertyCacheCreator: decouple resolution of property type and flags
This patch aims to improve readability and maintainability of the logic
for deriving property type and flags before adding it to cache.
Now resolution of the type happens in one place and derivation of flags
in another one. This not only improves separation of concerns, but
also makes the dependency between attributes and type more
explicit and hopefully easier to follow.

Task-number: QTBUG-98320
Change-Id: I64424e959ee11aa5c9a90c8e9fd2ca1347715342
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>