QtObject: Give uiLanguage a signal in addition to BINDABLE
The consensus at contributor summit [1] was that we want to phase out
the BINDABLE integration over time. That means that we shouldn't have
properties which are BINDABLE but non-NOTIFY-able in Qt itself exposed
to QML. Adjust QtObject::uiLanguage accordingly.

No revision marker is needed as the global Qt object only exists as a
singleton (and is ina  private header).

[1] https://wiki.qt.io/QtCS25_-_QProperty_and_QBindable

Fixes: QTBUG-138749
Pick-to: 6.10
Change-Id: Ice54f844c87864a325e667d2536c39af821ce7b9
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>