QSGThreadedRenderLoop: Handle window hiding at polish step
A window can be hidden during the polish step. For example, an app may
hide its window if the geometry is not good enough. This can result in
the following render thread events if the QPA sends sync expose events:

 - WM_Obscure
 - WM_TryRelease
 - WM_RequestSync

The WM_RequestSync event comes from the WMSyncEvent.

With QtWayland, repainting a window after it has gotten unexposed can
lead to the compositor posting a protocol error. There is also matter of
keeping code predictable: if the QPA sends an unexpose event, it'd be
great if the app stops repainting.

Pick-to: 6.10
Change-Id: I1d477436d6f79156aca44960c47a84f2e12ca4d2
Reviewed-by: Andy Nichols <andy.nichols@qt.io>