qmlls: find C++ definitions of properties
Implement go to definition for properties of QML components defined in
C++. Jump to the C++ file at the correct line when searching for
properties ofQML Component types from QML.

This allows user to jump to their own C++ files, and does not jump to
headers outside the project folder, like private/qquickitem_p.h for
example, as it might get confusing for the user to open a file in the
editor that they are not supposed to edit.

Also log a warning when a C++ header was not found.

Task-number: QTBUG-128393
Change-Id: I9da962b671dfe1f9e29fb260655f4f37132c6815
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>