tst_qmlls_utils: remove invalid test for attached derived type
We want to warn about using attached derived types in QTBUG-138873, so
remove the test that was testing whether go to definition jumps to the
derived type MyApplicationWindow when querying for the
MyApplicationWindow.footer attached property of the ApplicationWindow
base type.
A later commit will make qmlls jump to the C++ definition of the
attached type, instead of the QML type.

Task-number: QTBUG-128393
Change-Id: Ieb96e94315fe8f053302aee3f44cd2437c81fbe0
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>