qmllsutils: distinguish attached types in bindings
Attached types on the left hand side of a binding have a different
scopes than attached types on the right hand side of a binding.
Special case them in resolveExpressionType to allow to implement
go-to-definition on attached types correctly for both kind of scopes in
a later commit.

Task-number: QTBUG-128393
Change-Id: Id82f3572014c8c8310cc359c4ed0bbfb6465cf7c
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>