qmlls: find C++ definitions of grouped properties
Implement go to definition for grouped properties.
Jump to the C++ file at the correct line when searching for grouped
properties from QML.

This allows user to jump to their own C++ files, and does not jump to
headers outside the project folder, like private/qquickitem_p.h for
example, as it might get confusing for the user to open a file in the
editor that they are not supposed to edit.

Task-number: QTBUG-128393
Change-Id: I190a71b51b618e4050c9542f987d67e9b835a59f
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>