qmlls: find C++ definitions of attached types
Implement go to definition for attached types defined in C++.
Jump to the C++ file at the correct line when searching for attached
types or their methods and properties from QML.

Support both cases of attached types:
Attached types in the binding target property or in the binding itself.

This allows user to jump to their own C++ files, and does not jump to
headers outside the project folder, like private/qquickitem_p.h for
example, as it might get confusing for the user to open a file in the
editor that they are not supposed to edit.

Task-number: QTBUG-128393
Change-Id: I4c40d7241227d2e5921c85ab7756782a37e5a0ad
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>