QuickDialogs: Do not use QFLAGS
Dialogs uses QFLAGS to bring in various classes via related
meta-objects. The goal is to expose specific enums in the QML API.

What it does instead is to expose all enums of those classes at runtime,
but not to tooling.

Switch the code to use QML_EXTENDED_NAMESPACE, which also exposes all
enums at runtime, but also to tooling.

Task-number: QTBUG-137075
Pick-to: 6.10
Change-Id: Id8d5858575a445f957572fc090df482db1b6fce8
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>