Acknowlege touch event state within the popup as similar to mouse
The touch events ignored by the pop-up have not been updated with
the handled status, and this makes the delivery agent consider
the pop-up as the event grabber, which further blocks the touch
event from propagating.

This patch makes the pop-up to update the event with the touch
event handling status. Thus, even the ignored status has been
updated in the event, allowing the delivery agent to propagate
the events to other items.

Fixes: QTBUG-138478
Pick-to: 6.10 6.9 6.8 6.5
Change-Id: I2399aea70d58b6d427bf76f871a3fcc04c69a44f
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>