Run tst_qquickfiledialogimpl tests for multiple dialog styles
We were only testing the non-native file dialog under the Basic style
previously, but we have 4 other styles (Fusion, Imagine, Material,
Universal) that were never tested.

This unfortunately caused recent regressions in both the FileDialog and
FolderDialog.

By running each test once for more styles, we should be better
protected against future regressions in individual styles.

In this patch, we enable all tests for Basic, Fusion, Material and
Universal. The Imagine style is intentionally omitted due to a bug,
which causes the tst_qquickfiledialogimpl::defaults() test to fail.

This will be fixed later.

Change-Id: I9b4ab722c5ad3052122024ad3a7a2f0a99de77e6
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>