Adapt to removal of QAbstactFileEngine::entryList
The method is (going to be) removed in 6.11 and is currently producing a
compilation error, as the method is marked `override` but is not
overrinding anything anymore, on current development builds that use a
recent qtbase.

Follow along the base class path and the resolution method in similar
patches (e.g for Qt Creator) by removing the method in its entirety.

Change-Id: I6a334681764d9589317ef2121691f8d4e7888b98
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>