Use new qt_target_qml_from_svg() function for examples
This updates weatherforecast and vectorimage examples to
use the automatic svgtoqml build steps instead of manual
conversion.

Removes .pro file for weatherforecast. This is a new
example after the move to CMake and it should have
never had a .pro file to begin with. We aren't going
to implement the svgtoqml build step with qmake, so we
can't keep this working.

Task-number: QTBUG-128915
Change-Id: I2e2506316d4fc8fb35c227309f50e0c92f2a0e38
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>