tst_qmlformat_cli: remove hand-written keys
Remove the hand-written keys from the test and use the static strings
instead. One of the previous hand-written key had a typo. Also make sure
that the settings are really set, as QVariant::toBool() returns false
for a unset settings, for example in the case of the key with the typo.

Change-Id: I15da10409fbf374a475dc3930844fd56120f370a
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>