Fusion MenuSeparator: Make it stronger and wider in high contrast mode
The MenuSeparator used a subtle color based on the darkShade color from
the Fusion singleton.

In high contrast mode, we want it to have more contrast from the Menu's
background. And, while subjective, I personally think it could be wider
too, in order to emphesize the separation of MenuItems.

Task-number: QTBUG-134971
Pick-to: 6.10
Change-Id: Ifa3e178831dafbf8b1d74e00da8e81d733efbe33
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>