ScrollView: use padding if no rightPadding or bottomPadding is set
Because rightPadding and bottomPadding properties are set in the windows
and macos ScrollView style implementation if the user try to set the
padding property there will be no effect. So with this change we make
sure that the padding property is used if rightPadding or bottomPadding
aren't explicitly set by the user.

Fixes: QTBUG-123631
Change-Id: I5078cec1774cef7b99e0735022ddb4ce84cd08df
Pick-to: 6.10 6.8 6.7
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>