QtQml: Return QVariant{Map|Hash} as-is when converting QJSValue
If we know that the object contained in a QJSValue was created from a
QVariantMap or QVariantHash, it makes no sense to hide that by producing
the JavaScript representation of it. The user is calling toVariant() to
obtain a usable QVariant after all.

[ChangeLog][QtQml][Important Behavior Changes] QJSValue::toVariant() now
returns a QVariantMap or QVariantHash for JavaScript objects created
from a QVariantMap or QVariantHash, even if you tell it to
RetainJSObjects via the "behavior" argument. There is no conversion in
this case, after all.

Change-Id: I482cd1d2db310048de066aad33efbbe9c3f9e56a
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>