Compiler: Add support for multiple document edits per fix suggestion
The suggested fix to a warning may require multiple document edits
to properly fix the issue.

Imagine an unqualified lookup across a component boundary. One of the
fixes the user should be able to choose from is to qualify the lookup by
adding the id _and_ set 'pragma ComponentBehavior: Bound'. This fix
requires editing the document in two different places.

To this end, FixSuggestion no longer holds a single replacement string
directly but instead holds a list of, newly introduced, DocumentEdit.

Note that FixSuggestion itself still keeps a location as well. This is
intended to point out the location of the warning to the user, which may
differ from the places where the fixes are applied. And, this is also
needed in case we need to warn and suggest a fix but cannot do it
automatically for the user.

The fix suggesting to change 'new Array(1, 2)' into '[1, 2]' was updated
to use the new multiple edits, simplifying the logic.

The output of qmllint was updated to communicate the edits associated
with a fix to the user. The json format was also updated.

The qmllint integration in qmlls was also updated. Testing in creator
shows that multiple edits can be applied correctly for a single fix.

The QQmlSA API was also updated to include a DocumentEdit of its own and
to update FixSuggestion as a consequence. QQmlSA is still in tech
preview.

Change-Id: Ifedb9a50329f0f3e9e59a74981a51e0daab52ae9
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>