qqmlcodemodel: extract version checking code into static method
Prepare for QTBUG-119163 that will require some kind of "force update"
to reconstruct the DomItem once the CMake build in the background
finishes.
Extract the version-testing logic into static methods, one to test the
version of the "normal" document, one for the valid document, and one
that acts on the results of the two first methods.
This will allow to pass a "force" flag in the future that skips the
version checks.

Also add capitalization to the warnings and correct typos.

Task-number: QTBUG-119163
Change-Id: Ia72be615b527a571332640ceb6c7f6085f03ded9
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>