qqmlcodemodel: add UpdatePolicy enum for snapshot updates
Add an UpdatePolicy enum for snapshot updates to be able to force update
snapshots and force relint files, even if they have the same version.
Also replace an existing enum that was doing the same thing with
UpdatePolicy, and replace some methods that was accepting a 'force'
boolean argument to take an UpdatePolicy enum.

Task-number: QTBUG-119163
Change-Id: I4b7e2f4229a3aa5bccd798232dc9cfef04c88b9f
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>