mac style: fall back to draw ProgressBar manually on macOS Tahoe 26
On macOS Tahoe 26, the drawing of a ProgressBar using the mac
style is broken. Nothing is drawn if we continue to use
[NSView drawRect:], [NSView bitmapImageRepForCachingDisplayInRect:]
or [CALayer renderInContext:] (and similar API) for drawing
an NSProgressIndicator.

This patch will therefore implement a fall back approach where
we draw the ProgressBar manually when we detect that the current
version of macOS is running with liquid glass support.

Pick-to: 6.10 6.9 6.8 6.5
Task-number: QTBUG-138947
Task-number: QTBUG-138942
Change-Id: I26faea6e24d48bf0cb7bf2cd2b114d6ca2c2825f
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>