Doc: Unify cli options and file settings in a single table for qmlformat
The current documentation for the qmlformat tool uses two separate
tables to list the available command line options and the available
settings file variables.

Some of the content is duplicated between the two tables, as the
settings file variables always have an equivalent command line option to
configure the same behavior.

Despite the duplication, the two tables aren't entirely in sync and
don't necessarily provide the same amount of information.
For example, the command line options table misses some of the default
value for certain configurable behaviors or, similarly, certain
descriptions might be more thorough in one table rather than the other.

Since both tables generally should present the same information for
those features that can be configured both from a command line option
and from a settings file variable, as they present two different but
supposedly equivalent interfaces to the same feature, the tables are now
unified into one, so as to ensure that the description and default value
for the relevant elements can more easily remain in sync, further
reducing the unnecessary duplication between the tables.

In particular, the command line options table was expanded with an
additional column, `Setting Name`, that is set to "N/A" when the
command line option has no equivalent settings file variable and
contains the relevant variable name, taken from the other table, when it
does.

The default values for the relevant element were moved from the
settings file table to the now unified table where necessary.
Since the command line options use flags for boolean variables, where a
behavior is implicitly disabled unless the flag itself is present as
part of the invocation, the meaning of "a default value of false" could
be confusing, possibly suggesting that the user could pass a "false" or
"true" value to the flag to set the behavior.
To avoid this confusion, the "Default Value" column was modified to be a
"Default State/Value" column, and the cell for rows that are affected by
the above issues were modified to be defined as "disabled/false",
hopefully making it more explicit that there is a difference between the
state-based command line usage and the value-based settings file usage.

The description for elements of the above kind was extended to be more
specific about the difference in usage with regards to a command line
invocation and a settings file customization.

The descriptions for the relevant element were brought over from the
settings file table where they either provided additional information or
provided stronger layout for the presented information.

The original settings file table was removed as its information is now
embedded in the unified table.

A "\target qmlformat-settings" command that pointed to the now removed
settings file table was removed as it was unused in the project,
with the consideration that the unified table already provides a
"\target" that can be linked to if necessary.

The title of the section presenting the table was modified to "Options
and settings" to explicit the presence of the new unified information.

A phrase referring to the command line options table as presenting the
available command line options was modified to explicit the presence, in
the now unified table, of both command line options and settings file
variables.

Fixes: QTBUG-139125
Change-Id: Iedcc9ee0a1b3a70417db10736cef4a2955b43750
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>