Add QT_QML_DEBUG compile definition to the Qt6::Qml target
Consider the QT_ENABLE_QML_DEBUG target property when building
QML executables and add the QT_QML_DEBUG compiler definition to enable
QML debugging. The property is defined and intialized with the
QT_ENABLE_QML_DEBUG variable value on target creation. But can be set
manually.

TODO: Leave the flag undocumented in this commit. Since it's the public
API change, we only document it in further releases, but QtC may use it
with all supported Qt versions the undocumented "private" way.

Pick-to: 6.10 6.9 6.8
Task-number: QTBUG-139551
Change-Id: I94d4269981f87d6c226d7565257786658d68f992
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>