OutWriter: fix skipComment logic and rename some methods
Essentially itemStart and itemEnd, as well as RegionStart and RegionEnd
are dealing with Comments.
However if we set skipComment=true and remove itemStart, as for example
from the MethodInfo::signature, this will lead to crash because of the
top() call on empty pendingComments stack.

This commit fixes this issue, hopefully making this logic more explicit + does some renaming

Change-Id: Id2789ef4005534199d307affba785d2f697131a2
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>