VectorImage: Support basic masks
This implements support for SVG masks with the default content
unit type. The alternative "object bounding box" content unit type
is still not supported as it would need additional logic to replicate
the mask per masked item.

With this change, the parser has to provide a unique ID for each node,
so that references to masks can be resolved by the parser. This involves
some complexity in the SVG parser to avoid name collisions for use nodes,
switch nodes, etc.

Task-number: QTBUG-121537
Change-Id: If9f47f364f8682ab6021e7f6277c480c3f075297
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>