QQmlDelegateModel: De-bounce initialization of two-way bindings
When installing the reverse binding on a QQmlDelegateModelItem, the
binding is evaluated right away and tries to write the model data. We
don't want that since we know that the model data already contains the
same value.

Use a temporary filtering metaobject to block writes to this property
while installing the reverse binding. This causes the bounced write to
fail.

Amends commit 4bd5b312795024bd5bcd698525b58caf3893c913

Pick-to: 6.10
Task-number: QTBUG-132420
Change-Id: Ib001f669046635cc3ae830154f22d95ee1d3924c
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>