Attached ToolTip: show automatically if visible not set
This makes attached ToolTips show automatically if the attached visible
property hasn't been explicitly set:

    Button {
        ToolTip.text: qsTr("Blah")
    }

The tool tip will be shown on hover or long press (if triggered by
touch and the attachee is an AbstractButton).

As discussed in the linked issue, there are existing projects that
set visible imperatively, which means that doing this automatically
without an easy way to opt out would break code. For this reason, an
attached policy property is also introduced.

In addition to visibility, the delay and timeout properties are also
set automatically if they haven't been explicitly set. The attached
policy property also controls this behavior.

[ChangeLog][Controls][ToolTip] Added policy property that controls
whether attached ToolTips are shown automatically if the visible
property hasn't been explicitly set. The delay and timeout properties
are also set automatically if they weren't explicitly set.

Task-number: QTBUG-116910
Change-Id: I188b2d2f56c8834863c0714c0064073d48360e00
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>